// src/app/models/member.ts
import { Photo } from "./photo";

export interface Member {
    id: number;
    userName: string;
    age: number;
    photosUrl: string;
    knownAS: string;
    created: Date;
    lastActive: Date;
    gender: string;
    introduction: string;
    interests: string;
    lookingFor: string;
    city: string;
    country: string;
    photos: Photo[];
}